+++ src/Exception/InvalidMaxDepthException.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Torr\SimpleNormalizer\Exception;
+
+final class InvalidMaxDepthException extends \InvalidArgumentException implements NormalizerExceptionInterface
+{
+}